app/Port_App/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
requests
python-dotenv
google-generativeai